requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.11.1
pyyaml==6.0.1 